refactor: type slide over to be a stack of slides
