+++ apps/hash-frontend/src/pages/shared/entity-type-page.tsx
@@ -50,7 +50,7 @@ import { FileUploadsTab } from "./entity-type-page/file-uploads-tab";
 import { EntityTypeContext } from "./entity-type-page/shared/entity-type-context";
 import { EntityTypeHeader } from "./entity-type-page/shared/entity-type-header";
 import { useCurrentTab } from "./entity-type-page/shared/tabs";
-import { TypePreviewSlide } from "./entity-type-page/type-preview-slide";
+import { TypeSlideOverStack } from "./entity-type-page/type-slide-over-stack";
 import { useEntityTypeValue } from "./entity-type-page/use-entity-type-value";
 import { TopContextBar } from "./top-context-bar";
 
@@ -422,11 +422,9 @@ export const EntityTypePage = ({
       </EntityTypeFormProvider>
 
       {previewEntityTypeUrl ? (
-        <TypePreviewSlide
-          key={previewEntityTypeUrl}
+        <TypeSlideOverStack
+          rootTypeId={previewEntityTypeUrl}
           onClose={() => setPreviewEntityTypeUrl(null)}
-          onNavigateToType={onNavigateToType}
-          typeUrl={previewEntityTypeUrl}
         />
       ) : null}
